+++ app/src/main/java/io/legado/app/ui/book/read/page/entities/TextPoint.kt
@@ -2,5 +2,5 @@ package io.legado.app.ui.book.read.page.entities
 
 data class TextPoint(
     val x: Float,
-    val y: Int
+    val y: Float
 )
